제작 : windowCtlogMaptool - 맵툴용 카탈로그 윈도우
추가 : windowMvCatalog
- 스크롤 화면내 범위

+++ project/directBase/begin/labelBase.cpp
@@ -15,7 +15,7 @@ labelBase::~labelBase()
 
 void labelBase::drawDoUI(void)
 {
-	RECT rcText = { _position.x , _position.y, 0, 0 };
+	RECT rcText = { (LONG)_position.x , (LONG)_position.y, 0, 0 };
 
 	_font->DrawTextA(
 		NULL,
